new post controllers with authentication integrated

+++ app/Http/Controllers/Api/V1/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,7 +16,10 @@ class PostController extends Controller
      */
     public function index()
     {
-        return PostResource::collection(Post::with("author")->paginate());
+        $user = request()->user();
+        $posts = $user->posts()->paginate();
+        // return PostResource::collection($posts);
+        return $posts;
     }
 
     /**
@@ -24,7 +28,7 @@ class PostController extends Controller
     public function store(StorePostRequest $request)
     {
         $data = $request->validated();
-        $data["author_id"] = 1;
+        $data["author_id"] = $request->user()->id;
 
         $post = Post::create($data);
         return new PostResource($post);
@@ -35,6 +39,12 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        // $user = request()->user();
+        // if ($user->id != $post->author->id) {
+        //     abort(403, "Access forbidden!");
+        // }
+
+        abort_if(Auth::id() != $post->author_id, 403, "Access Forbidden!");
         return new PostResource($post);
     }
 
@@ -43,8 +53,9 @@ class PostController extends Controller
      */
     public function update(StorePostRequest $request, Post $post)
     {
-        $data = $request->validated();
+        abort_if(Auth::id() != $post->author_id, 403, "Access Forbidden!");
 
+        $data = $request->validated();
         $post->update($data);
         return new PostResource($post);
     }
@@ -54,6 +65,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Auth::id() != $post->author_id, 403, "Access Forbidden!");
+
         $post->delete();
         return response()->noContent();
     }
